pyproject.toml
[project]
name = "ecoscout"
version = "0.1.0"
description = "EcoScout -- Live Environmental Intelligence Companion for the Gemini Live Agent Challenge"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
authors = [
    { name = "EcoScout Team" }
]

dependencies = [
    "google-adk>=1.18.0",
    "google-genai>=1.0.0",
    "google-cloud-firestore>=2.19.0",
    "google-cloud-storage>=2.18.0",
    "fastapi>=0.115.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "uvicorn[standard]>=0.32.0",
    "pyinaturalist>=0.21.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "slow: marks tests that require API calls (deselect with '-m \"not slow\"')",
]

[tool.uv]
dev-dependencies = []

[tool.hatch.build.targets.wheel]
packages = ["app"]
